+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Palette Changelog
 > A palette fieldtype built for [Craft 3](http://craftcms.com)
 
+## [Unreleased]
+
+### Added
+
+*   Added preset settings logic
+
 ## 1.0.1 - 2018-09-04
 
 ### Added